requirements.txt
beautifulsoup4
requests
urllib3
pyyaml